requirements.txt
numpy==2.3.4
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==12.0.0

-e scripts/py360convert
